Move useFocusItemsHelper
