+++ generatedTypes/hooks/useFocusItemsHelper/index.d.ts
@@ -30,7 +30,7 @@ export declare type Props = {
      */
     addOffsetMargin?: boolean;
     /**
-     * How much space (padding \ margin) is there on he left\right of the items
+     * How much space (padding \ margin) is there on the left\right of the items
      */
     outerSpacing?: number;
     /**
